pyproject.toml
[tool.poetry]
name = "lms"
version = "0.1.0"
description = "Ruslana.English - LMS system"
authors = ["s.kravtsov <s.kravtsov@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.3"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^2024.2"
psycopg2 = "^2.9.9"
django-ckeditor-5 = "^0.2.13"
django-modeltranslation = "^0.19.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
